changed phone number type from int to long

+++ healthrecords-backend/src/main/java/spring/demo/healthrecordsbackend/controller/HealthrecordsBackendController.java
@@ -27,7 +27,7 @@ public class HealthrecordsBackendController {
                                   String dob,
                                   String doctorFirstName,
                                   String doctorLastName,
-                                  int doctorPhoneNumber,
+                                  long doctorPhoneNumber,
                                   boolean allergies,
                                   boolean diabetes,
                                   boolean highBloodPressure){
